fix(types): handle optional campaignId prop for CampaignCard and Content components
This commit updates the prop type definitions for the CampaignCard and CampaignCard/Content components to properly handle the optional campaignId prop. Done by defining types AdminProps and NonAdminProps and using a union of these two types in the component props definitions. 

This (hopefully) allows for correct interpretation of the campaignId prop based on whether or not isAdmin is true. It fixes the type error that was previously occurring.