feat(frontend): Implement redesign for admin page (#447)
* WIP: feat(admin): use CampaignCard

* feat(fe): Replace children prop with isAdmin for conditional rendering of popover in CampaignCard

* feat(fe): Create initial dropdown and dropdown option components and add to admin campaign card

* feat(fe): add @headlessui/talwindcss to frontend

* feat(fe): Implement edit and delete functionality by passing appropriate props to dropdown option component

* feat(fe): Remove unnecessary logs

* refactor(fe): Move modal and related state/callbacks from AdminCampaignComponent to CampaignCard/Content

* fix(types): handle optional campaignId prop for CampaignCard and Content components

This commit updates the prop type definitions for the CampaignCard and CampaignCard/Content components to properly handle the optional campaignId prop. Done by defining types AdminProps and NonAdminProps and using a union of these two types in the component props definitions. 

This (hopefully) allows for correct interpretation of the campaignId prop based on whether or not isAdmin is true. It fixes the type error that was previously occurring.

---------

Co-authored-by: Michael Vo <zax@zaxu.xyz>